+++ src/App.css
@@ -7,6 +7,4 @@
     position: relative;
     
     /* top: 30px; */
-    
-
 }
