+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import React from 'react';
 import Header from './Components/Header/Header';
-import ProductContainer from './Components/Product/productContainer';
+import ProductContainer from '../src/Pages/HomePage/HomePage'
 import { CartProvider } from './contexts/CartContext';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import CartPage from './Pages/CartPAge/CartPage';
@@ -10,13 +10,18 @@ const App: React.FC = () => {
   return (
     <div className="App">
       <Router>
+
         <CartProvider> 
+
           <Header />
+          
           <Routes>
             <Route path="/" element={<ProductContainer />} />
             <Route path="/cart" element={<CartPage />}  />
           </Routes>
+
         </CartProvider>
+
       </Router>
     </div>
   );
