--- src/Components/Product/productContainer.css
@@ -1,101 +1,0 @@
-* {
-   box-sizing: border-box;
- }
- 
-.container {
-   display: flex;
-   flex-direction: column;
-   margin-top: 7%;
- }
- 
-.row {
-   margin: auto;
- }
- 
-.card-custom {
-   width: 15rem;
-   height: 380px;
-   background-color: rgb(248, 238, 208) !important;
-   margin-top: 8% !important;
-   display: flex;
-   flex-direction: column;
-   align-items: center;
-   padding: 1rem;
-   overflow: hidden;
- }
- 
- .banner-img {
-   width: 100%;
-   height: 350px;
- }
-
- .heading {
-   text-align: center;
-   padding-top: 30px;
-   margin-bottom: 1rem;
- }
- 
-
- .card-img-custom {
-   height: 120px;
-   width: 190px;
-   margin: 10px;
- }
- 
-
- .card-body-custom {
-   width: 100%;
-   display: flex;
-   flex-direction: column;
-   align-items: center;
-   text-align: center;
-   overflow: hidden;
- }
- 
-
- .card-title {
-   overflow: hidden;
-   text-overflow: ellipsis;
-   display: -webkit-box;
-   -webkit-line-clamp: 2;
-   -webkit-box-orient: vertical;
-   margin-bottom: 0.5rem;
- }
- 
-
- .button-card {
-   display: block;
-   margin: 10px 0; 
-   width: 100%;
-   box-sizing: border-box;
- }
-
- .button-card.hidden {
-   display: none;
- }
- 
- .btn-custom {
-   --bs-btn-padding-y: .25rem;
-   --bs-btn-padding-x: .5rem;
-   --bs-btn-font-size: .75rem;
- }
- 
-
- .container-cust {
-   display: flex;
-   justify-content: center;
-   align-items: center;
-   margin-top: 1rem;
- }
- 
-
- .container-cust.hidden {
-   display: none;
- }
- 
- .btn-custom-container {
-   display: flex;
-   justify-content: center;
-   gap: 0.5rem; 
- }
-
