+++ src/Components/ProductCard/ProductCard.css
@@ -1,0 +1,36 @@
+.card-custom {
+    width: 15rem;
+    height: 380px;
+    background-color: rgb(248, 238, 208) !important;
+    margin-top: 8% !important;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 1rem;
+    overflow: hidden;
+  }
+
+  .card-img-custom {
+    height: 120px;
+    width: 190px;
+    margin: 10px;
+  }
+  
+ 
+  .card-body-custom {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+    overflow: hidden;
+  }
+  
+  .card-title {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+    margin-bottom: 0.5rem;
+  }
